Streamlit App/requirements.txt
absl-py==0.11.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==1.5.0
blinker==1.4
boto3==1.17.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dlib==19.21.1
entrypoints==0.3
gast==0.4.0
gitdb==4.0.5
GitPython==3.1.12
grpcio==1.35.0
html5lib==0.9999999
icc-rt==2019.0
idna==2.10
importlib-metadata==3.4.0
imutils==0.5.4
intel-openmp==2021.1.2
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
olefile==0.46
opencv-python-headless
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywinpty==0.5.7
pyzmq==22.0.2
requests==2.25.1
scipy==1.1.0
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
streamlit==0.75.0
tbb==2021.1.1
tbb4py==2021.1.1
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.4.0
